@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -27,23 +27,23 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.0
-  velocity_x: ^1.0.1
-  firebase_core: ^0.5.2+1
-  firebase_auth: ^0.18.3+1
-  firebase_database: ^4.3.0
-  connectivity: ^2.0.2
-  google_maps_flutter: ^1.0.6
-  geolocator: ^6.1.10
-  http: ^0.12.2
-  provider: ^4.3.2+2
-  flutter_polyline_points: ^0.2.4
-  font_awesome_flutter: ^8.10.1
-  animated_text_kit: ^3.1.0
+  cupertino_icons: ^1.0.3
+  velocity_x: ^3.0.0
+  firebase_core: ^1.2.0
+  firebase_auth: ^1.2.0
+  firebase_database: ^7.0.0-dev.3
+  connectivity: ^3.0.4
+  google_maps_flutter: ^2.0.5
+  geolocator: ^7.0.3
+  http: ^0.13.3
+  provider: ^5.0.0
+  flutter_polyline_points: ^1.0.0
+  font_awesome_flutter: ^9.0.0
+  animated_text_kit: ^4.2.1
 
-  flutter_bloc: ^6.1.0
-  equatable: ^1.2.5
-  dartz: ^0.9.2
+  flutter_bloc: ^7.0.0
+  equatable: ^2.0.2
+  dartz: ^0.10.0-nullsafety.2
 
 
 
