@@ -1,9 +1,9 @@
 class Address {
-  String placeName;
-  double latitude;
-  double longitude;
-  String formattedPlaceAddress;
-  String placeID;
+  String? placeName;
+  double? latitude;
+  double? longitude;
+  String? formattedPlaceAddress;
+  String? placeID;
   Address({
     this.placeName,
     this.latitude,
