@@ -1,10 +1,10 @@
 import 'package:firebase_database/firebase_database.dart';
 
 class UserModel {
-  String id;
-  String fullName;
-  String phone;
-  String email;
+  String? id;
+  String? fullName;
+  String? phone;
+  String? email;
 
   UserModel({this.id, this.email, this.phone, this.fullName});
 
