@@ -5,14 +5,14 @@ import 'package:velocity_x/velocity_x.dart';
 class TaxiButton extends StatelessWidget {
   final String buttonText;
   final Function onPressed;
-  final Color color;
+  final Color? color;
 
-  TaxiButton({@required this.onPressed, @required this.buttonText, this.color});
+  TaxiButton({required this.onPressed, required this.buttonText, this.color});
 
   @override
   Widget build(BuildContext context) {
     return RaisedButton(
-      onPressed: onPressed,
+      onPressed: onPressed as void Function()?,
       shape: RoundedRectangleBorder(
         borderRadius: BorderRadius.circular(25),
       ),
