@@ -5,7 +5,7 @@ import 'package:http/http.dart' as http;
 class RequestHelper {
   static Future<dynamic> getRequest(String url) async {
     try {
-      http.Response response = await http.get(url);
+      http.Response response = await http.get(Uri.parse(url));
       if (response.statusCode == 200) {
         // success
         String data = response.body;
